// src/types.ts
export interface components {
    schemas: {
// ...
        categoryObject: {
            /**
             * Format: int32
             * @description A system defined unique identifier for the category
             */
            id: number;
            /** @description The name of the category */
            name: string;
            /** @description The description of the category or `null` if not set */
            description: string | null;
            /** @description If `true`, the transactions in this category will be treated as income. (See <a href="https://support.lunchmoney.app/setup/categories/category-properties">Category Properties</a> for more details) */
            is_income: boolean;
            /** @description If `true`, the transactions in this category will be excluded from the budget. (See <a href="https://support.lunchmoney.app/setup/categories/category-properties">Category Properties</a> for more details) */
            exclude_from_budget: boolean;
            /** @description If `true`, the transactions in this category will be excluded from totals. (See <a href="https://support.lunchmoney.app/setup/categories/category-properties">Category Properties</a> for more details) */
            exclude_from_totals: boolean;
            /**
             * Format: date-time
             * @description The date and time of when the category was last updated (in the ISO 8601 extended format).
             */
            updated_at: string;
            /**
             * Format: date-time
             * @description The date and time of when the category was created (in the ISO 8601 extended format).
             */
            created_at: string;
            /**
             * Format: int64
             * @description The ID of the category group this category belongs to or `null` if the category doesn't belong to a group, or is itself a category group.
             */
            group_id: number | null;
            /** @description If `true`, the category is created as a category group */
            is_group: boolean;
            /** @description For category groups, this will populate with details about the categories that belong to this group. The objects in this array are similar to Category Objects but do not include the `is_income`, `exclude_from_budget`, and `exclude_from_totals` properties as these are inherited from the category group. In addition, the `is_group` property will always be `false`, and there will be no `children` attribute. */
            children?: components["schemas"]["childCategoryObject"][];
            /** @description If true, the category is archived and not displayed in relevant areas of the Lunch Money app. */
            archived: boolean;
            /**
             * Format: date-time
             * @description The date and time of when the category was last archived (in the ISO 8601 extended format).
             */
            archived_at: string | null;
            /**
             * @description An integer specifying the position in which the category is displayed on the categories page in the Lunch Money GUI. For categories within a category group the order is relative to the other categories within the group.<br>Categories with `order: null` will be displayed in alphabetical order by name, prior to any categories with an order
             * @default null
             */
            order: number | null;
            /**
             * @description If `true`, the category is collapsed in the Lunch Money GUI
             * @default false
             */
            collapsed: boolean;
        };
// ...
        manualAccountObject: {
            /**
             * Format: int32
             * @description The unique identifier of this account
             */
            id: number;
            /** @description Name of the account */
            name: string;
            /** @description Name of institution holding the account */
            institution_name: string | null;
            /** @description Optional display name for the account as set by the user or derived from the `institution_name` and `name` if not explicitly set. */
            display_name: string | null;
            /** @description Primary type of the account */
            type: components["schemas"]["accountTypeEnum"];
            /** @description Optional account subtype. Examples include<br> - retirement - checking - savings - prepaid credit card */
            subtype: string | null;
            /** @description Current balance of the account in numeric format to 4 decimal places */
            balance: string;
            /** @description Three-letter lowercase currency code of the account balance */
            currency: string;
            /** @description The balance converted to the user's primary currency */
            to_base: number;
            /**
             * Format: date-time
             * @description Date balance was last updated in ISO 8601 extended format, can be in date or date-time format
             */
            balance_as_of: string;
            /**
             * @description The status of the account
             * @enum {string}
             */
            status: "active" | "closed";
            /**
             * Format: date
             * @description The date this account was closed in YYYY-MM-DD format. Will be null if the account has not been marked as closed.
             */
            closed_on: string | null;
            /** @description An optional external_id that may be set or updated via the API */
            external_id: string | null;
            /** @description User defined JSON data that can be set or cleared via the API */
            custom_metadata?: {
                [key: string]: unknown;
            } | null;
            /**
             * @description If true, this account will not show up as an option for assignment when creating transactions manually
             * @default false
             */
            exclude_from_transactions: boolean;
            /** @description The name of the user who created the account */
            created_by_name: string;
            /**
             * Format: date-time
             * @description Date/time the account was created in ISO 8601 extended format
             */
            created_at: string;
            /**
             * Format: date-time
             * @description Date/time the account was created in ISO 8601 extended format
             */
            updated_at: string;
        };
// ...
        tagObject: {
            /**
             * Format: int32
             * @description Unique identifier for the tag
             */
            id: number;
            /** @description Name of the tag */
            name: string;
            /** @description Description of the tag */
            description: string | null;
            /** @description The text color of the tag */
            text_color: string | null;
            /** @description The background color of the tag */
            background_color: string | null;
            /**
             * Format: date-time
             * @description The date and time of when the tag was last updated (in the ISO 8601 extended format).
             */
            updated_at: string;
            /**
             * Format: date-time
             * @description The date and time of when the tag was created (in the ISO 8601 extended format).
             */
            created_at: string;
            /** @description If `true`, the tag will not show up when creating or updating transactions in the Lunch Money app. **Can it be assigned via the API** */
            archived: boolean;
            /**
             * Format: date-time
             * @description The date and time of when the tag was last archived or `null` if not archived
             */
            archived_at: string | null;
        };
// ...
    };
// ...
}